Dataprocess/requirements.txt
pymongo>=4.3.2
load_dotenv>=0.1.0
pandas>=1.5.2 
matplotlib>=3.5.2
ipykernel>=6.19.4